requirements.txt
pbr
ruamel.yaml
